move detr3d_transformer.py to /layers
