+++ projects/detr3d/mmdet3d_plugin/__init__.py
@@ -5,6 +5,6 @@ from .models.detectors.detr3d import Detr3D
 from .models.task_modules.hungarian_assigner_3d import HungarianAssigner3D
 from .models.task_modules.match_cost import BBox3DL1Cost
 from .models.task_modules.nms_free_coder import NMSFreeCoder
-from .models.utils.detr3d_transformer import (Detr3DCrossAtten,
-                                              Detr3DTransformer,
-                                              Detr3DTransformerDecoder)
+from .models.layers.detr3d_transformer import (Detr3DCrossAtten,
+                                               Detr3DTransformer,
+                                               Detr3DTransformerDecoder)
